requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.3.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
